Moving circle toggle button to props!
This is hard work...
