+++ App/client/src/App.js
@@ -2,8 +2,9 @@ import React, { Component } from 'react';
 import * as d3 from 'd3';
 import axios from 'axios';
 // import Toolbox from './Toolbox';
-import CustomOverlay from './CustomOverlay';
+import CircleButton from './Components/CircleButton';
 import './App.less';
+import Circle from './CustomOverlay/Circle';
 
 class App extends Component {
     constructor(props) {
@@ -20,9 +21,7 @@ class App extends Component {
             leftClick: undefined, // Will set state to leftClick listener
             rightClick: undefined, // Will set state to rightClick listener
 
-            toggleColor: true,
-
-            mouseEvent: undefined // Will set mouse event here from listener
+            toggleColor: true
         };
 
     }
@@ -41,64 +40,64 @@ class App extends Component {
         this.mainPageLoad(map);
     }
 
-    drawingComponent = () => {
-        let startPos;
-        const naver = window.naver;
-        const { map } = this.state;
-        const { circleToggle } = this.state;
+    // drawingComponent = () => {
+    //     let startPos;
+    //     const naver = window.naver;
+    //     const { map } = this.state;
+    //     const { circleToggle } = this.state;
 
-        if (circleToggle === true) {
-            const leftClick = naver.maps.Event.addListener(map, 'click', e => {
-                // coord: lat, lng of map
-                // offset: x, y of screen
-                const { coord, offset } = e;
-                startPos = { coord, offset };
+    //     if (circleToggle === true) {
+    //         const leftClick = naver.maps.Event.addListener(map, 'click', e => {
+    //             // coord: lat, lng of map
+    //             // offset: x, y of screen
+    //             const { coord, offset } = e;
+    //             startPos = { coord, offset };
 
-                naver.maps.Event.removeListener(leftClick);
-            });
+    //             naver.maps.Event.removeListener(leftClick);
+    //         });
 
-            const rightClick = naver.maps.Event.addListener(map, 'rightclick', e => {
-                const { coord, offset } = e;
-                const endPos = { coord, offset };
-                new CustomOverlay({
-                    position: { startPos, endPos },
-                    naverMap: map,
-                    zoom: ''
-                }).setMap(map);
+    //         const rightClick = naver.maps.Event.addListener(map, 'rightclick', e => {
+    //             const { coord, offset } = e;
+    //             const endPos = { coord, offset };
+    //             new Circle({
+    //                 position: { startPos, endPos },
+    //                 naverMap: map,
+    //                 zoom: ''
+    //             }).setMap(map);
 
-                naver.maps.Event.removeListener(rightClick);
-            });
+    //             naver.maps.Event.removeListener(rightClick);
+    //         });
 
-            this.setState({ rightClick: rightClick });
-            this.setState({ leftClick: leftClick });
-        }
-        this.setState({ circleToggle: !circleToggle }); // Complete shape and turn off toggle
-    };
+    //         this.setState({ rightClick: rightClick });
+    //         this.setState({ leftClick: leftClick });
+    //     }
+    //     this.setState({ circleToggle: !circleToggle }); // Complete shape and turn off toggle
+    // };
 
-    ellipseState() {
-        const { circleToggle } = this.state;
-        this.setState({ circleToggle: !circleToggle });
-    }
+    // ellipseState() {
+    //     const { circleToggle } = this.state;
+    //     this.setState({ circleToggle: !circleToggle });
+    // }
 
-    removeListener() {
-        const naver = window.naver;
-        const { leftClick } = this.state;
-        const { rightClick } = this.state;
-        naver.maps.Event.removeListener(leftClick);
-        naver.maps.Event.removeListener(rightClick);
-    }
+    // removeListener() {
+    //     const naver = window.naver;
+    //     const { leftClick } = this.state;
+    //     const { rightClick } = this.state;
+    //     naver.maps.Event.removeListener(leftClick);
+    //     naver.maps.Event.removeListener(rightClick);
+    // }
 
-    changeColor() {
-        const { toggleColor } = this.state;
-        this.setState({ toggleColor: !toggleColor });
-    }
+    // changeColor() {
+    //     const { toggleColor } = this.state;
+    //     this.setState({ toggleColor: !toggleColor });
+    // }
 
-    circleToggleAndEllipseAndChangeColor = () => {
-        this.changeColor();
-        this.drawingComponent();
-        this.ellipseState();
-        this.removeListener();
-    }
+    // circleToggleAndEllipseAndChangeColor = () => {
+    //     this.changeColor();
+    //     this.drawingComponent();
+    //     this.ellipseState();
+    //     this.removeListener();
+    // }
 
     mapOption = () => {
         const naver = window.naver;
@@ -138,7 +137,7 @@ class App extends Component {
                         const { startPos, endPos, zoomLevel } = JSON.parse(
                             el.figures
                         );
-                        return new CustomOverlay({
+                        return new Circle({
                             position: { startPos, endPos },
                             naverMap: map,
                             zoom: zoomLevel
@@ -170,14 +169,16 @@ class App extends Component {
     }
 
     render() {
-        const { toggleColor } = this.state;
-        const btnClass = toggleColor ? 'lightPurple' : 'darkPurple';
+        const { map } = this.state;
+        // const { toggleColor } = this.state;
+        // const btnClass = toggleColor ? 'lightPurple' : 'darkPurple';
         return (
             <div id="wrapper">
                 <div id="map" onClick={this.mouseClick} onContextMenu={this.mouseClick} onKeyDown={this.mouseClick}>
                     {/* <Toolbox mapLoad={mapLoad} /> */}
                 </div>
-                <button type="button" className={btnClass} onClick={this.circleToggleAndEllipseAndChangeColor}>Circle</button>
+                {/* <button type="button" className={btnClass} onClick={this.circleToggleAndEllipseAndChangeColor}>Circle</button> */}
+                <CircleButton nmap={map} />
             </div>
         );
     }
